gjør om til ts og bruker typer
